requirements.txt
autopep8==1.4.4
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-autoslug-iplweb==1.9.4.dev0
django-countries==5.3.3
django-model-utils==3.1.2
django-phonenumber-field==3.0.1
html5lib==1.0.1
idna==2.8
oauthlib==3.0.1
phonenumbers==8.10.13
Pillow==6.0.0
progressbar2==3.42.0
psycopg2==2.8.3
pycodestyle==2.5.0
PyPDF2==1.26.0
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.1
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.0
tqdm==4.32.2
Unidecode==1.1.0
urllib3==1.25.3
webencodings==0.5.1
xhtml2pdf==0.2.3
